+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/KeyAgreementPublicValue.kt
@@ -3,6 +3,7 @@ package at.asitplus.signum.indispensable
 import at.asitplus.awesn1.Asn1Encodable
 import at.asitplus.awesn1.Asn1Sequence
 import at.asitplus.awesn1.Asn1Decodable
+import at.asitplus.signum.indispensable.sign.ECDSAPublicKey
 
 /**
  * Key agreement public value. Must be PEM encodable/decodable.
@@ -13,9 +14,9 @@ interface KeyAgreementPublicValue : Asn1Encodable<Asn1Sequence> {
      */
     interface ECDH: KeyAgreementPublicValue {
         /**
-         * Returns this value ad a [CryptoPublicKey.EC]
+         * Returns this value as an [ECDSAPublicKey]
          */
-        fun asCryptoPublicKey(): CryptoPublicKey.EC
+        fun asCryptoPublicKey(): ECDSAPublicKey
     }
     companion object : Asn1Decodable<Asn1Sequence, ECDH> {
         override fun doDecode(src: Asn1Sequence) = CryptoPublicKey.doDecode(src) as CryptoPublicKey.EC